+++ src/main/java/duckoboye/femboymod/lists/ToolMaterialList.java
@@ -1,8 +1,9 @@
-package duckoboye.nickelmod.lists;
+package duckoboye.femboymod.lists;
 
 import net.minecraft.item.IItemTier;
 import net.minecraft.item.Item;
 import net.minecraft.item.crafting.Ingredient;
+
 // An enum that set the properties for a material (in this case the Nickel one)
 public enum ToolMaterialList implements IItemTier {
     // We create the nickel sword properties here (the basic ones)
@@ -11,6 +12,7 @@ public enum ToolMaterialList implements IItemTier {
     private float attackDamage, efficiency;
     private int durability, harvestLevel, enchantability;
     private Item repairMaterial;
+
     // We init all the attributes in the constructor
     private ToolMaterialList(float attackDamage, float efficiency, int durability, int harvestLevel, int
             enchantability, Item repairMaterial) {
@@ -21,16 +23,34 @@ public enum ToolMaterialList implements IItemTier {
         this.enchantability = enchantability;
         this.repairMaterial = repairMaterial;
     }
+
     @Override
-    public float getAttackDamage() { return this.attackDamage; }
+    public float getAttackDamage() {
+        return this.attackDamage;
+    }
+
     @Override
-    public float getEfficiency() { return this.efficiency; }
+    public float getEfficiency() {
+        return this.efficiency;
+    }
+
     @Override
-    public int getEnchantability() { return this.enchantability; }
+    public int getEnchantability() {
+        return this.enchantability;
+    }
+
     @Override
-    public int getHarvestLevel() { return this.harvestLevel;}
+    public int getHarvestLevel() {
+        return this.harvestLevel;
+    }
+
     @Override
-    public int getMaxUses(){ return this.durability; }
+    public int getMaxUses() {
+        return this.durability;
+    }
+
     @Override
-    public Ingredient getRepairMaterial() { return Ingredient.fromItems(this.repairMaterial); }
+    public Ingredient getRepairMaterial() {
+        return Ingredient.fromItems(this.repairMaterial);
+    }
 }
