+++ src/main/resources/assets/femboymod/lang/en_us.json
@@ -1,0 +1,4 @@
+{
+  "item.femboymod.nickel": "Nickel",
+  "item.femboymod.femboy_sword": "Femboy \"Sword\""
+}
