+++ src/main/resources/assets/femboymod/models/item/femboy_sword.json
@@ -1,6 +1,6 @@
 {
   "parent": "item/handheld",
   "textures": {
-    "layer0": "nickelmod:item/femboy_sword"
+    "layer0": "femboymod:item/femboy_sword"
   }
 }
