+++ src/main/resources/assets/femboymod/models/item/nickel.json
@@ -1,6 +1,6 @@
 {
   "parent": "item/generated",
   "textures": {
-    "layer0": "nickelmod:item/nickel"
+    "layer0": "femboymod:item/nickel"
   }
 }
