+++ src/main/resources/assets/femboymod/textures/item/femboy_sword_old.png
