+++ src/main/resources/assets/femboymod/textures/item/nickel.png
