--- src/main/resources/assets/nickelmod/lang/en_us.json
@@ -1,4 +1,0 @@
-{
-  "item.nickelmod.nickel": "Nickel",
-  "item.nickelmod.femboy_sword": "Femboy \"Sword\""
-}
